[#122] ref: pipe_utils.c: norm OK
