+++ include/prototype.h
@@ -6,7 +6,7 @@
 /*   By: yuotsubo <yuotsubo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 14:29:07 by yuotsubo          #+#    #+#             */
-/*   Updated: 2024/12/13 17:57:08 by yuotsubo         ###   ########.fr       */
+/*   Updated: 2024/12/13 15:37:32 by tkitahar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,6 +76,9 @@ void	reset_redirect(t_node *redir);
 void	prepare_pipe_parent(t_node *node);
 void	err_exit(const char *name, const char *err_msg, int estatus);
 char	**token_list_to_argv(t_token *tok);
+char	*dup_until_space(const char *s);
+void	validate_access(const char *path, const char *filename);
+size_t	argv_len(t_token *tok);
 
 // parse
 t_node	*pipeline(t_token **rest, t_token *tok);
